requirements.txt
asyncpg
aiogram~=3.15.0
APScheduler~=3.10.4
python-decouple
SQLAlchemy~=2.0.36